Rename BlogItem file for consistency
